Move windows specific code utilities to their own module
